+++ src/molecules/GasReceivedChart/GasReceivedChart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Skeleton from '../../atoms/Skeleton/Skeleton';
+import SkeletonChart from '../../molecules/SkeletonChart/SkeletonChart';
 import Message from '../../atoms/Message/Message';
 
 import { useQuery, gql } from '@apollo/client';
@@ -23,9 +23,8 @@ const GAS_EXCHANGES_CHART = gql`
 	}
 `;
 
-const now = DateTime.now();
-
 export default function GasReceivedChart() {
+	const now = DateTime.now();
 	const { loading, error, data } = useQuery(GAS_EXCHANGES_CHART, {
 		variables: {
 			resolution: 'TWO_HOURS',
@@ -36,7 +35,7 @@ export default function GasReceivedChart() {
 		}
 	});
 
-	if (loading) return <Skeleton width="100%" height="300px" />;
+	if (loading) return <SkeletonChart />;
 	if (error) return <Message type="error">{error.message}</Message>;
 
 	return (